requirements.txt
streamlit>=1.30.0
discord.py>=2.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
